Renamed namespaces for Integration.Vsix Analysis and CFamily files
* no code changes
* fixed up namespaces for corresponding tests
* moved CLangAnalyzerTests to the correct subfolder
