+++ src/Integration.Vsix.UnitTests/Analysis/AnalysisConfigMonitorTests.cs
@@ -23,9 +23,9 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using SonarLint.VisualStudio.Core;
 using SonarLint.VisualStudio.Integration.NewConnectedMode;
-using SonarLint.VisualStudio.Integration.Vsix.Analysis;
+using SonarLint.VisualStudio.Integration.UnitTests;
 
-namespace SonarLint.VisualStudio.Integration.UnitTests
+namespace SonarLint.VisualStudio.Integration.Vsix.Analysis.UnitTests
 {
     [TestClass]
     public class AnalysisConfigMonitorTests
