+++ src/Integration.Vsix.UnitTests/Analysis/AnalysisRequesterTests.cs
@@ -21,9 +21,9 @@
 using System;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SonarLint.VisualStudio.Integration.Vsix.Analysis;
+using SonarLint.VisualStudio.Integration.UnitTests;
 
-namespace SonarLint.VisualStudio.Integration.UnitTests
+namespace SonarLint.VisualStudio.Integration.Vsix.Analysis.UnitTests
 {
     [TestClass]
     public class AnalysisRequesterTests
