+++ src/Integration.Vsix.UnitTests/Analysis/DisableRuleCommandTests.cs
@@ -30,10 +30,10 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using SonarLint.VisualStudio.Core;
 using SonarLint.VisualStudio.Integration.NewConnectedMode;
-using SonarLint.VisualStudio.Integration.Vsix;
-using SonarLint.VisualStudio.Integration.Vsix.Analysis;
+using SonarLint.VisualStudio.Integration.UnitTests;
+using ThreadHelper = SonarLint.VisualStudio.Integration.UnitTests.ThreadHelper;
 
-namespace SonarLint.VisualStudio.Integration.UnitTests.SonarLintDaemon
+namespace SonarLint.VisualStudio.Integration.Vsix.Analysis.UnitTests
 {
     [TestClass]
     public class DisableRuleCommandTests
