+++ src/Integration.Vsix.UnitTests/Analysis/SingleFileMonitorTests.cs
@@ -25,9 +25,9 @@ using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using SonarLint.VisualStudio.Core.SystemAbstractions;
-using SonarLint.VisualStudio.Integration.Vsix;
+using SonarLint.VisualStudio.Integration.UnitTests;
 
-namespace SonarLint.VisualStudio.Integration.UnitTests
+namespace SonarLint.VisualStudio.Integration.Vsix.Analysis.UnitTests
 {
     [TestClass]
     public class SingleFileMonitorTests
