+++ src/Integration.Vsix.UnitTests/Analysis/StatusBarReanalysisProgressHandlerTests.cs
@@ -21,9 +21,9 @@
 using System;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SonarLint.VisualStudio.Integration.Vsix;
+using SonarLint.VisualStudio.Integration.UnitTests;
 
-namespace SonarLint.VisualStudio.Integration.UnitTests
+namespace SonarLint.VisualStudio.Integration.Vsix.Analysis.UnitTests
 {
     [TestClass]
     public class StatusBarReanalysisProgressHandlerTests
