+++ src/Integration.Vsix/Analysis/StatusBarReanalysisProgressHandler.cs
@@ -22,9 +22,8 @@ using System;
 using System.Diagnostics;
 using Microsoft.VisualStudio;
 using Microsoft.VisualStudio.Shell.Interop;
-using SonarLint.VisualStudio.Integration.Vsix.Analysis;
 
-namespace SonarLint.VisualStudio.Integration.Vsix
+namespace SonarLint.VisualStudio.Integration.Vsix.Analysis
 {
     /// <summary>
     /// Handles displaying reanalysis progress in the VS status bar
