+++ src/Integration.Vsix/Analysis/UserSettingsProvider.cs
@@ -24,9 +24,8 @@ using System.Diagnostics;
 using System.IO;
 using SonarLint.VisualStudio.Core;
 using SonarLint.VisualStudio.Core.SystemAbstractions;
-using SonarLint.VisualStudio.Integration.Vsix.Analysis;
 
-namespace SonarLint.VisualStudio.Integration.Vsix
+namespace SonarLint.VisualStudio.Integration.Vsix.Analysis
 {
     [Export(typeof(IUserSettingsProvider))]
     internal sealed class UserSettingsProvider : IUserSettingsProvider, IDisposable
